force to int data_amount
force to int data_amount to use it in template with stripe checkout button:
    <script
        src="https://checkout.stripe.com/checkout.js" class="stripe-button"
        data-key="{{ form.stripeToken.field.widget.attrs.data_key }}"
        data-image="{{ STATIC_URL }}img/logo.png"
        data-name="Site Name"
        data-description="{{ form.stripeToken.field.widget.attrs.data_description }}"
        data-amount="{{ form.stripeToken.field.widget.attrs.data_amount }}"
        data-currency="{{ form.stripeToken.field.widget.attrs.data_currency }}">
    </script>

+++ payments/stripe/widgets.py
@@ -13,7 +13,7 @@ class StripeWidget(HiddenInput):
             'data_key': provider.public_key,
             'data_description': payment.description or _('Total payment'),
             # Stripe accepts cents
-            'data_amount': payment.total * 100,
+            'data_amount': int(payment.total * 100),
             'data_currency': payment.currency
         }
         kwargs['attrs'].update(attrs)
